requirements.txt
accelerate==1.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
cssselect==1.3.0
exceptiongroup==1.3.0
faiss-cpu
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.0
h11==0.16.0
hf-xet==1.1.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.4
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.5.0
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.9.1
numpy==1.26.4
pillow==11.2.1
psutil==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
safetensors==0.5.3
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==3.2.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sympy==1.14.0
threadpoolctl==3.6.0
tinysegmenter==0.3
tldextract==5.3.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.45.2
typing_extensions==4.13.2
urllib3==2.4.0
